+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -1,12 +1,12 @@
 package codegen
 
 object FlutterXPluginInfo {
-    const val Version: String = "5.3.1.243"
+    const val Version: String = "5.3.2.242"
     const val Changelog: String = """
-<h2>5.3.1 - 2025-03-03</h2>
+<h2>5.3.2 - 2025-03-04</h2>
 
-<ul><li>Fix: dart function comment generator tool error (dart comment)</li><li>Fix: dart assets hover document preview bug (with image) (dart assets preview)</li><li>New: Supports the operation of clicking on the asset path to quickly open files（Need to hold ctrl）</li></ul>
+<ul><li>New: <code>flutter android gradle</code> Migration Tools (android tool)</li><li>Fix: <code>dio</code> The problem of listening server not being destroyed (dio tool)</li></ul>
 """
 
-    const val BuildDate: String = "2025-03-03"
+    const val BuildDate: String = "2025-03-04"
 }